+++ server/src/main/java/org/allaymc/server/command/defaults/OpCommand.java
@@ -1,10 +1,8 @@
 package org.allaymc.server.command.defaults;
 
 import org.allaymc.api.command.tree.CommandTree;
-import org.allaymc.api.entity.interfaces.EntityPlayer;
 import org.allaymc.api.message.TrKeys;
-
-import java.util.Collection;
+import org.allaymc.api.server.Server;
 
 /**
  * @author daoge_cmd
@@ -17,28 +15,20 @@ public class OpCommand extends VanillaCommand {
 
     @Override
     public void prepareCommandTree(CommandTree tree) {
-        tree.getRoot().playerTarget("player").exec(context -> {
-            Collection<EntityPlayer> players = context.getResult(0);
-            if (players.isEmpty()) {
-                context.addPlayerNotFoundError();
+        tree.getRoot().wildcardTarget("player").exec(context -> {
+            String player = context.getResult(0);
+
+            var manager = Server.getInstance().getPlayerManager();
+            if (manager.isOperator(player)) {
+                context.addError("%" + TrKeys.MC_COMMANDS_OP_FAILED, player);
                 return context.fail();
             }
 
-            if (players.size() > 1) {
-                context.addTooManyTargetsError();
-                return context.fail();
-            }
-
-            var player = players.stream().findFirst().get();
-            if (player.isOperator()) {
-                context.addError("%" + TrKeys.MC_COMMANDS_OP_FAILED, player.getOriginName());
-                return context.fail();
-            }
-
-            player.setOperator(true);
-
-            context.addOutput(TrKeys.MC_COMMANDS_OP_SUCCESS, player.getOriginName());
-            player.sendTranslatable(TrKeys.MC_COMMANDS_OP_MESSAGE);
+            manager.setOperator(player, true);
+            context.addOutput(TrKeys.MC_COMMANDS_OP_SUCCESS, player);
+            manager.getPlayers().values().stream()
+                    .filter(p -> p.getLoginData().getUuid().toString().equals(player) || p.getOriginName().equals(player))
+                    .forEach(p -> p.sendTranslatable(TrKeys.MC_COMMANDS_OP_MESSAGE));
             return context.success();
         });
     }