+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Unless otherwise specified, any version comparison below is the comparison of th
 - (API) `Permission` object now can have multiple attached listeners.
 - (API) `PermissionGroup` can now override the permissions of the parent permission group to `false`.
 - (API) Added `resetPermission` method to `PermissionGroup` and `Permissible` which would set the permission status to `default`.
+- (API) Added methods `PlayerManager.isOperator()` and `PlayerManager.setOperaotor()`.
 - Command nodes that the player doesn't have permission to access will not be sent to the player, and the commands will
   be resent when the specific permission changes.
 - Updated log4j2 patterns to use `%logger{0}` instead of `%c{0}`.
@@ -28,6 +29,8 @@ Unless otherwise specified, any version comparison below is the comparison of th
 ### Changed
 
 - (API) Replaced method `Permission.getPermissionListener()` with `Permission.getListeners()` which returns a set of listeners.
+- Permission data now will not be saved into the player's nbt data, which allows third permission plugins to store the permission data in their own way.
+  The core will only saved a `operator` list now.
 
 ### Fixed